requirements.txt

python-telegram-bot>=20.0
requests>=2.28.0
python-whois>=0.8.0
dnspython>=2.3.0
ipaddress
cryptography>=3.4.0
urllib3>=1.26.0
python-dotenv>=0.19.0
python-whois
Flask>=3.0.0
Flask-Cors>=4.0.0
